Timer Project - Practice and Real-Life Application
Timer Project - Practice and Real-Life Application of JavaScript:

This project is a simple timer application created using HTML, CSS, and JavaScript. It was built as a practice project to apply JavaScript skills in real-life situations, focusing on timing functions and user interaction. The goal is to showcase how core web technologies can be used to create interactive and functional web apps.